# [AIRFLOW-6856] Bulk fetch paused_dag_ids