--- a/airflow/dag/base_dag.py
+++ b/airflow/dag/base_dag.py
@@ -63,14 +63,6 @@ class BaseDag(metaclass=ABCMeta):
         raise NotImplementedError()
 
     @abstractmethod
-    def is_paused(self):
-        """
-        :return: whether this DAG is paused or not
-        :rtype: bool
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
     def pickle_id(self):
         """
         :return: The pickle ID for this DAG, if it has one. Otherwise None.